modified:   src/Console/InstallCommand.php

+++ src/Console/InstallCommand.php
@@ -574,8 +574,23 @@ EOF;
     {
 
         if (file_exists(base_path('database/factories/CompanyFactory.php'))) {
+            if (file_exists(base_path('database/migrations/2020_05_21_100000_create_companies_table.php'))) {
             unlink(base_path('database/factories/CompanyFactory.php'));
-            unlink(base_path('database/migrations/*.php'));
+            $base_path = base_path('database/migrations');
+            $files = glob($base_path . '/*');
+            foreach ($files as $file) {
+                if (is_file($file)) {
+                    unlink($file);
+                }}
+            }
+            $this->callSilent('vendor:publish', ['--tag' => 'fortify-migrations', '--force' => true]);
+            // Sanctum...
+            (new Process([$this->phpBinary(), 'artisan', 'vendor:publish', '--provider=Laravel\Sanctum\SanctumServiceProvider', '--force'], base_path()))
+            ->setTimeout(null)
+            ->run(function ($type, $output) {
+                $this->output->write($output);
+            });
+            $this->configureSession();
         }
 
         // Publish Team Migrations...
@@ -632,8 +647,23 @@ EOF;
     {
 
         if (file_exists(base_path('database/factories/TeamFactory.php'))) {
+            if (file_exists(base_path('database/migrations/2020_05_21_100000_create_teams_table.php'))) {
             unlink(base_path('database/factories/TeamFactory.php'));
-            unlink(base_path('database/migrations/*.php'));
+            $base_path = base_path('database/migrations');
+            $files = glob($base_path . '/*');
+            foreach ($files as $file) {
+                if (is_file($file)) {
+                    unlink($file);
+                }}
+            }
+            $this->callSilent('vendor:publish', ['--tag' => 'fortify-migrations', '--force' => true]);
+            // Sanctum...
+            (new Process([$this->phpBinary(), 'artisan', 'vendor:publish', '--provider=Laravel\Sanctum\SanctumServiceProvider', '--force'], base_path()))
+            ->setTimeout(null)
+            ->run(function ($type, $output) {
+                $this->output->write($output);
+            });
+            $this->configureSession();
         }
 
         // Publish Team Migrations...
